requirements.txt
httplib2==0.7.4
oauth2==1.5.211
python-twitter==0.8.2
simplejson==2.5.0
wsgiref==0.1.2
